+++ plugin/inc/db.class.php
@@ -88,4 +88,122 @@ class PluginIserviceDB extends DB
         return true;
     }
 
+    public static function createTable(string $tableName, array $tableConfig, ?\DBmysql $db = null): bool
+    {
+        if ($db === null) {
+            global $DB;
+            $db = $DB;
+        }
+
+        if ($db->tableExists($tableName, false)) {
+            return self::alterTable($tableName, $tableConfig, $db);
+        }
+
+        $query  = "create table $tableName (";
+        $query .= implode(
+            ', ', [
+                self::getColumnsCreateModifySql($tableName, $tableConfig, $db),
+                self::getIndexesCreateModifySql($tableName, $tableConfig, $db),
+            ]
+        );
+        $query  = rtrim($query, ', ');
+        $query .= ')';
+
+        return $db->query($query) === true;
+    }
+
+    public static function alterTable(string $tableName, array $tableConfig, ?\DBmysql $db = null): bool
+    {
+        if ($db === null) {
+            global $DB;
+            $db = $DB;
+        }
+
+        if (!$db->tableExists($tableName, false)) {
+            return self::createTable($tableName, $tableConfig, $db);
+        }
+
+        $query  = "alter table $tableName";
+        $query .= implode(
+            ', ', [
+                self::getColumnsCreateModifySql($tableName, $tableConfig, $db),
+                self::getIndexesCreateModifySql($tableName, $tableConfig, $db),
+            ]
+        );
+        $query  = rtrim($query, ', ');
+
+        return $db->query($query) === true;
+    }
+
+    public static function getColumnsCreateModifySql(string $tableName, array $tableConfig, ?\DBmysql $db = null): string
+    {
+        if (empty($tableConfig['columns'])) {
+            return '';
+        }
+
+        if ($db === null) {
+            global $DB;
+            $db = $DB;
+        }
+
+        $sql    = '';
+        $action = '';
+
+        foreach ($tableConfig['columns'] as $columnName => $columnConfig) {
+            if ($db->tableExists($tableName)) {
+                $action = $db->fieldExists($tableName, $columnName) ? ' modify column' : ' add column';
+            }
+
+            $sql .= "$action `$columnName` $columnConfig,";
+        }
+
+        return rtrim($sql, ',');
+    }
+
+    public static function getIndexesCreateModifySql(string $tableName, array $tableConfig, ?\DBmysql $db = null): string
+    {
+        if (empty($tableConfig['indexes'])) {
+            return '';
+        }
+
+        if ($db === null) {
+            global $DB;
+            $db = $DB;
+        }
+
+        $sql    = '';
+        $action = '';
+
+        foreach ($tableConfig['indexes'] as $indexConfig) {
+            if ($db->tableExists($tableName)) {
+                $action = 'add';
+
+                if (self::indexExists($tableName, $indexConfig['name'], $db)) {
+                    continue;
+                }
+            }
+
+            $sql .= "$action $indexConfig[type] $indexConfig[name] $indexConfig[columns],";
+        }
+
+        return rtrim($sql, ',');
+    }
+
+    public static function indexExists(string $tableName, string $indexName, ?\DBmysql $db = null): bool
+    {
+        if ($db === null) {
+            global $DB;
+            $db = $DB;
+        }
+
+        if ($indexName == 'primary key') {
+            $indexName = 'primary';
+        }
+
+        $query  = "show index from $tableName where key_name = '$indexName'";
+        $result = $db->query($query);
+
+        return $result && $db->numrows($result) > 0;
+    }
+
 }