+++ plugin/install/sql/delete_tables.sql
@@ -1,21 +1,45 @@
-drop table if exists `glpi_plugin_iservice_configs`;
+drop table if exists `glpi_plugin_iservice_pendingemails`;
 
-drop table if exists `glpi_plugin_iservice_importmappings`;
+drop table if exists `glpi_plugin_iservice_downloads`;
 
-drop table if exists `hmarfa_facrind`;
+drop table if exists `glpi_plugin_iservice_movements`;
 
-drop table if exists `hmarfa_facturi`;
+drop table if exists `glpi_plugin_iservice_minimum_stocks`;
 
-drop table if exists `hmarfa_firme`;
+drop table if exists `glpi_plugin_iservice_consumables_models`;
 
-drop table if exists `hmarfa_gestiuni`;
+drop table if exists `glpi_plugin_iservice_orderstatuschanges`;
 
-drop table if exists `hmarfa_incpla`;
+drop table if exists `glpi_plugin_iservice_extorders`;
 
-drop table if exists `hmarfa_lotm`;
+drop table if exists `glpi_plugin_iservice_ememails`;
 
-drop table if exists `hmarfa_nommarfa`;
+drop table if exists `glpi_plugin_iservice_orderstatuses`;
+
+drop table if exists `glpi_plugin_iservice_intorders_extorders`;
+
+drop table if exists `glpi_plugin_iservice_intorders`;
+
+drop table if exists `glpi_plugin_iservice_cartridges_tickets`;
+
+drop table if exists `glpi_plugin_iservice_consumables_tickets`;
 
 drop table if exists `hmarfa_tran`;
 
-drop view if exists hmarfa_total_facturi;
+drop table if exists `hmarfa_nommarfa`;
+
+drop table if exists `hmarfa_lotm`;
+
+drop table if exists `hmarfa_incpla`;
+
+drop table if exists `hmarfa_gestiuni`;
+
+drop table if exists `hmarfa_firme`;
+
+drop table if exists `hmarfa_facturi`;
+
+drop table if exists `hmarfa_facrind`;
+
+drop table if exists `glpi_plugin_iservice_importmappings`;
+
+drop table if exists `glpi_plugin_iservice_configs`;