+++ plugin/install/sql/seed_database.sql
@@ -1,0 +1,2 @@
+insert into `glpi_plugin_iservice_configs` (`name`, `value`)
+values ('version', '0');